Spanel datatype changes from int to uint16_t

+++ src/gen_cranberry/gc_dev_spanel.cpp
@@ -4,13 +4,13 @@ void gc_dev_spanel_open(void){
     pinMode(_PIN_GC_SPANEL_, INPUT);
 }
 
-int gc_dev_spanel_read_raw(void){
-    int val = analogRead(_PIN_GC_SPANEL_);
+uint16_t gc_dev_spanel_read_raw(void){
+    uint16_t val = analogRead(_PIN_GC_SPANEL_);
     return val;
 }
 
-int gc_dev_spanel_read(void){
+uint16_t gc_dev_spanel_read(void){
     float raw = (float)analogRead(_PIN_GC_SPANEL_) * (5.0/1023.0) * 2.0;
-    int val = raw * 1000;
+    uint16_t val = raw * 1000;
     return val;
 }
